Updated Menu Progression
